% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_kmers.R
\name{find_kmers}
\alias{find_kmers}
\alias{find_manyKmers}
\alias{find_runsOfNs}
\title{Generic internal functions used by genespace}
\usage{
find_manyKmers(dnass, kmers, nCores = 1)

find_runsOfNs(dnass, minRunLength)
}
\arguments{
\item{dnass}{DNAStringSet}

\item{kmers}{DNAStringSet with kmers to find}

\item{nCores}{integer, then number of cores to use}

\item{bed}{data.table or data.frame with at least chr, start, end columns}

\item{seqInfo}{seqInfo ranges extracted from dnass}

\item{gr}{granges object}

\item{existingGr}{existing granges to mask with}

\item{newGr}{new granges to mask}

\item{x}{single-value parameter, string, integer, numeric, list, vector}

\item{...}{additional arguments passed on
\cr
If called, \code{find_kmers} returns its own arguments.}
}
\description{
\code{find_kmers} Convenience functions for gscTools, not meant to be called
directly by the user. Little documentation support provided, use at your own
risk.

\code{find_manyKmers} Finds exact DNA-DNA matches and aggregates across all
kmers in the kmer DNAstring

\code{find_runsOfNs} Exact string matching of Ns to a DNAStringSet to find
the position of gaps between contigs
}
