% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strandBam.R
\name{strandBam}
\alias{strandBam}
\title{Parse and plot bam files}
\usage{
strandBam(gff.file, bam.files, genes2test, min.dist = 120, buffer = 100,
  plotit = T, verbose = T, geneID.start = 4, geneID.end = 17,
  plot.pdf = "geneModel_reads.pdf")
}
\arguments{
\item{gff.file}{Character string indicating the name and file location of the
gff3 formatted annotation file}

\item{bam.files}{Character vector of the name(s) of the bam files to parse and concatenate. If the filese
are not in the working directory, the file location must also be included in
each bam.files name}

\item{genes2test}{Character string of the genes to examine. These must match the
ids in the gff file. See below}

\item{min.dist}{The minimum distance (bp) between any two genes. If all genes (regardless
whether they overlap) are desired, set to NA}

\item{buffer}{The bp distance around a gene model to plot and pull reads}

\item{plotit}{Logical, should plots be made and printed to a pdf?}

\item{verbose}{Logical, should updates be printed?}

\item{geneID.start}{The position of the first character of the gene id specified in
the gff3 file. This is typically 4, because the leading 3 characters are "ID=".}

\item{geneID.end}{The position of the last character of the gene id. Depends on
the format of the gene IDs. Typically a good idea to include both the gene identifier
and the splice variant so that all gene models are unique.}

\item{plot.pdf}{The name and file location of the output pdf (if plotit)}
}
\value{
A dataframe with the positions and flags of reads relating to each gene
}
\description{
\code{strandBam} Parses bam file(s) by the position of genes of interest, then
categorize the reads by their flags and plot them against the gene model
}
\details{
Here, we do some stuff. More here soon.
}
\examples{
# more here soon

}
